fix(ios): Update Swift code for Cordova 8.0+ compatibility
- Replace deprecated CDVCommandStatus_OK with .ok enum
- Replace deprecated CDVCommandStatus_ERROR with .error enum
- Fix force unwrap on non-optional CDVPluginResult
