+++ package.json
@@ -1,21 +1,22 @@
 {
   "name": "community-cordova-plugin-dbmeter",
-  "version": "2.2.0",
-  "description": "Cordova DBMeter plugin",
-  "author": {
-    "name": "Alexis Kofman"
-  },
-  "license": "Apache-2.0",
-  "cordova": {
-    "id": "community-cordova-plugin-dbmeter",
-    "platforms": [
-      "ios",
-      "android"
-    ]
-  },
+  "version": "2.3.1",
+  "main": "./types/index.d.ts",
+  "types": "./types/index.d.ts",
+  "readme": "README.md",
   "repository": "EYALIN/community-cordova-plugin-dbmeter/tree/master",
   "homepage": "https://github.com/EYALIN/community-cordova-plugin-dbmeter/tree/master",
   "funding": "https://github.com/sponsors/EYALIN",
+  "bugs": "https://github.com/EYALIN/community-cordova-plugin-dbmeter/issues",
+  "cordova": {
+    "id": "community-cordova-plugin-dbmeter",
+    "platforms": [
+      "android",
+      "ios"
+    ]
+  },
+  "description": "A Cordova plugin to get decibel (dB) values from the device microphone",
+  "license": "MIT",
   "keywords": [
     "cordova",
     "ios",
@@ -25,6 +26,8 @@
     "volume",
     "sound",
     "decibel",
+    "microphone",
+    "noise",
     "ecosystem:cordova",
     "cordova-ios",
     "cordova-android",
